+++ projects/frontend/es-app/src/guis/menus/GuiMenuGameSettings.h
@@ -38,6 +38,7 @@ class GuiMenuGameSettings : public GuiMenuBase
       QuitTwice,
       IntegerScale,
       Shaders,
+      Softpatching,
       ShaderSet,
       RetroAchivements,
       Netplay,
@@ -54,6 +55,8 @@ class GuiMenuGameSettings : public GuiMenuBase
     std::shared_ptr<SwitchComponent> mSmooth;
     //! Rewind
     std::shared_ptr<SwitchComponent> mRewind;
+    //! Shaders
+    std::shared_ptr<OptionListComponent<std::string>> mSoftpatching;
     //! Auto-save
     std::shared_ptr<SwitchComponent> mAutoSave;
     //! Press twice to quit
@@ -69,6 +72,8 @@ class GuiMenuGameSettings : public GuiMenuBase
     static std::vector<ListEntry<std::string>> GetRatioEntries();
     //! Get Shaders List
     static std::vector<ListEntry<std::string>> GetShadersEntries();
+    //! Get Softpatching List
+    static std::vector<ListEntry<std::string>> GetSoftpatchingEntries();
     //! Get shader sets List
     static std::vector<ListEntry<std::string>> GetShaderPresetsEntries();